requirements.txt
pymycobot
pyzmq
opencv-python
mediapipe